% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_data-data.R
\docType{data}
\name{basics_data}
\alias{basics_data}
\title{Basics Practice Data}
\format{
An object of class \code{data.frame} with 200 rows and 5 columns.
}
\usage{
data(basics_data)
}
\description{
This dataset is used in one of our tutorials,
please see the tutorial for more details.
}
\details{
- ID: an ID number for each business.
- Where: where the jobs were outsourced to.
- Jobs: the percent of outsourced jobs for
a call center.
- Cost: one calculation of the cost savings
for the business.
- Cost2: a separate way to calculate cost savings
for the business.
}
\keyword{datasets}
