% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cohens_d}
\alias{cohens_d}
\title{Calculate Effect Size (Cohen's d)}
\usage{
cohens_d(x, y, paired = FALSE)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{paired}{Whether data is paired}
}
\value{
Cohen's d
}
\description{
Calculates Cohen's d for paired or independent samples.
}
