% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-data.R
\docType{data}
\name{correlation}
\alias{correlation}
\title{Correlation Practice Data}
\format{
An object of class \code{data.frame} with 300 rows and 5 columns.
}
\usage{
data(correlation)
}
\description{
This dataset is used in one of our tutorials,
please see the tutorial for more details.
}
\details{
-	Gender of the participant surveyed on
these topics
-	Temporality: an average score of the rated
ability to adapt to change over time 1 (not
changing) to 7 (changing a lot)
-	Expectability: a rated degree of satisfaction
with the BI
-	Relativity: average score rating of how much
better one system is than another in BI 1 (not
very good) to 7 (very good)
-	Positive emotion: how positive participants
felt about BI (higher scores are more positive,
ranges from 1 to 7).
}
\keyword{datasets}
