% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascreen_data-data.R
\docType{data}
\name{datascreen_data}
\alias{datascreen_data}
\title{Data Screening Practice Data}
\format{
An object of class \code{data.frame} with 400 rows and 5 columns.
}
\usage{
data(datascreen_data)
}
\description{
This dataset is used in one of our tutorials,
please see the tutorial for more details.
}
\details{
a) Gender (1 = male, 2 = female)
b) Group (1 = control group, 2 =
experimental group)
c) 3 satisfaction scores, ranging from
2-125 points. Decimals are possible!
The control group was measured at the
same three time points, but did not take
part in the educational program.
   i) Before the program
   ii)	Half way through the program
   iii) After the program
}
\keyword{datasets}
