% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_data}
\alias{.validate_data}
\title{Validate Input Data}
\usage{
.validate_data(x, y, paired = FALSE)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{paired}{Whether data should be paired}
}
\value{
Logical indicating if data is valid
}
\description{
Checks if input vectors are valid for statistical tests.
}
\keyword{internal}
