% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-data.R
\docType{data}
\name{final}
\alias{final}
\title{Final Data}
\format{
An object of class \code{data.frame} with 60 rows and 5 columns.
}
\usage{
data(final)
}
\description{
This dataset is used as a template for our
final exam.
}
\keyword{datasets}
