% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p-value for Display}
\usage{
format_p_value(p, digits = 3)
}
\arguments{
\item{p}{p-value}

\item{digits}{Number of digits}
}
\value{
Formatted p-value string
}
\description{
Formats p-value according to common statistical conventions.
}
