% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midterm-data.R
\docType{data}
\name{midterm}
\alias{midterm}
\title{Midterm Data}
\format{
An object of class \code{data.frame} with 158 rows and 22 columns.
}
\usage{
data(midterm)
}
\description{
This dataset is used as a template for our
midterm exam.
}
\keyword{datasets}
