% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midterm_data.R
\name{midterm_data}
\alias{midterm_data}
\title{Exam Data Set Up}
\usage{
midterm_data(idnum = NULL)
}
\arguments{
\item{idnum}{Student ID number to set up data.}
}
\value{
midterm dataset
}
\description{
This function grabs the midterm/final data,
randomizes it for each student,
and allows them to each have a unique dataset.
}
\keyword{answers,}
\keyword{datasets}
\keyword{student}
