% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcmdr.R
\name{Rcmdr}
\alias{Rcmdr}
\title{Load R Commander
This way, run Rcmdr() and the environment will be ready.}
\usage{
Rcmdr()
}
\description{
Load R Commander
This way, run Rcmdr() and the environment will be ready.
}
\examples{
\dontrun{Rcmdr()}
}

