% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_data-data.R
\docType{data}
\name{regress_data}
\alias{regress_data}
\title{Regression Practice Data}
\format{
An object of class \code{data.frame} with 160 rows and 5 columns.
}
\usage{
data(regress_data)
}
\description{
This dataset is used in one of our tutorials,
please see the tutorial for more details.
}
\details{
- Dependent variable (Y): OCB - Organizational
citizenship behavior measure
-	Independent variables (X)
- Affective - job satisfaction measures that
measure emotion
- Cognitive - job satisfaction measures that
measure cognition (thinking)
-	Years - years on the job
-	Type_work - type of employee measured
(secretary, assistant, manager, boss)
}
\keyword{datasets}
