% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/learnstats.R
\docType{package}
\name{stats250}
\alias{stats250}
\alias{stats250-package}
\title{learnstats: Using R as an interactive educational environment.}
\description{
learnstats: Using R as an interactive educational environment.
}

