% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_data-data.R
\docType{data}
\name{ttest_data}
\alias{ttest_data}
\title{t-Test Practice Data}
\format{
An object of class \code{data.frame} with 200 rows and 3 columns.
}
\usage{
data(ttest_data)
}
\description{
This dataset is used in one of our tutorials,
please see the tutorial for more details.
}
\details{
-	Gender: male and female subjects were
examined in this experiment.
- PAL_cell: average physical activity values
for the cell phone accelerometer (range 0-100).
-	PAL_acc: average physical activity values
for the hand held accelerometer (range 0-100).
}
\keyword{datasets}
